Disable Spec Lifecycle Methods and Reorganize Files
